remove in-process endpoint probing in favor of OTel httpcheck
