+++ cmd/modelsrv/server.go
@@ -8,19 +8,14 @@ import (
 	"path/filepath"
 	"runtime"
 	"syscall"
-	"time"
 
 	"github.com/spf13/cobra"
 	"go.emeland.io/modelsrv/pkg/authz"
 	"go.emeland.io/modelsrv/pkg/backend"
 	"go.emeland.io/modelsrv/pkg/endpoint"
-	"go.emeland.io/modelsrv/pkg/endpointprobe"
-	"go.emeland.io/modelsrv/pkg/eventfilter"
 	"go.emeland.io/modelsrv/pkg/filesensor"
 	"go.uber.org/zap"
 )
-
-const certprobeShutdownTimeout = 30 * time.Second
 
 var serviceAddr string
 var dataDir string
@@ -29,12 +24,6 @@ var trustAuthHeaders bool
 var auditorIdentity string
 var auditorGroup string
 var publicResourceTypes string
-var enableCertprobe bool
-var certprobeInterval time.Duration
-var certprobeDebounce time.Duration
-var certprobeTimeout time.Duration
-var certprobeExpiryWarning time.Duration
-var maxConcurrentProbes int
 
 // serverCmd represents the server command
 var serverCmd = &cobra.Command{
@@ -69,7 +58,6 @@ var serverCmd = &cobra.Command{
 		logger.Infow("starting modelsrv",
 			"listen", serviceAddr,
 			"dataDir", dataPath,
-			"certprobe", enableCertprobe,
 		)
 		logger.Infof("REST API: http://%s/api", serviceAddr)
 		logger.Infof("Swagger UI: http://%s/swagger/", serviceAddr)
@@ -100,42 +88,6 @@ var serverCmd = &cobra.Command{
 			}
 		}
 
-		var scheduler *endpointprobe.Scheduler
-		var rescanFilterID eventfilter.FilterID
-		if enableCertprobe {
-			reg := endpoint.MetricsRegistry()
-			if reg == nil {
-				logger.Error("metrics registry not initialized; cannot start certprobe")
-				return
-			}
-
-			scheduler, err = endpointprobe.NewScheduler(endpointprobe.SchedulerConfig{
-				Client:              endpointprobe.NewModelClient(b.GetModel()),
-				Prober:              endpointprobe.NewProber(certprobeTimeout),
-				Metrics:             endpointprobe.NewMetrics(reg),
-				Interval:            certprobeInterval,
-				Debounce:            certprobeDebounce,
-				ExpiryWarning:       certprobeExpiryWarning,
-				MaxConcurrentProbes: maxConcurrentProbes,
-				Logger:              logger,
-				Publisher:           endpointprobe.NewModelFindingPublisher(b.GetModel(), logger),
-			})
-			if err != nil {
-				logger.Errorw("invalid certprobe configuration", "error", err)
-				return
-			}
-
-			go scheduler.Run(ctx)
-			rescanFilterID = b.GetChain().RegisterFilter(endpointprobe.NewRescanFilter(scheduler))
-			logger.Infow("certprobe started",
-				"interval", certprobeInterval,
-				"debounce", certprobeDebounce,
-				"timeout", certprobeTimeout,
-				"expiryWarning", certprobeExpiryWarning,
-				"maxConcurrentProbes", maxConcurrentProbes,
-			)
-		}
-
 		logger.Info("server is running (Ctrl+C to stop)")
 
 		sigCh := make(chan os.Signal, 1)
@@ -144,26 +96,6 @@ var serverCmd = &cobra.Command{
 		logger.Infow("shutdown signal received", "signal", sig.String())
 
 		cancel()
-
-		if scheduler != nil {
-			b.GetChain().Unregister(rescanFilterID)
-
-			shutdownCtx, shutdownCancel := context.WithTimeout(context.Background(), certprobeShutdownTimeout)
-			defer shutdownCancel()
-
-			done := make(chan struct{})
-			go func() {
-				scheduler.Wait()
-				close(done)
-			}()
-
-			select {
-			case <-done:
-				logger.Info("certprobe stopped")
-			case <-shutdownCtx.Done():
-				logger.Warn("shutdown timeout waiting for in-flight cert probes")
-			}
-		}
 
 		endpoint.StopWebListener()
 		logger.Info("goodbye")
@@ -180,26 +112,11 @@ func init() {
 	serverCmd.Flags().StringVar(&auditorIdentity, "auditor-identity", envOrDefault("AUDITOR_IDENTITY", ""), "OIDC subject treated as auditor when matching X-Auth-Subject")
 	serverCmd.Flags().StringVar(&auditorGroup, "auditor-group", envOrDefault("AUDITOR_GROUP", ""), "Group id treated as auditor when present in X-Auth-Groups")
 	serverCmd.Flags().StringVar(&publicResourceTypes, "public-resource-types", envOrDefault("PUBLIC_RESOURCE_TYPES", ""), "Comma-separated resource types always visible (e.g. ContextType,FindingType)")
-	serverCmd.Flags().BoolVar(&enableCertprobe, "enable-certprobe", envOrDefault("ENABLE_CERTPROBE", "true") == "true", "Run certificate probing as a background process inside modelsrv")
-	serverCmd.Flags().DurationVar(&certprobeInterval, "certprobe-interval", 5*time.Minute, "Certprobe background scan interval")
-	serverCmd.Flags().DurationVar(&certprobeDebounce, "certprobe-debounce", envDurationOrDefault("CERTPROBE_DEBOUNCE", 5*time.Second), "Debounce window before event-triggered certprobe rescans")
-	serverCmd.Flags().DurationVar(&certprobeTimeout, "certprobe-timeout", 10*time.Second, "Per-probe HTTP/TLS timeout")
-	serverCmd.Flags().DurationVar(&certprobeExpiryWarning, "expiry-warning", 720*time.Hour, "Raise CertificateExpiringSoon when remaining cert lifetime is at or below this duration")
-	serverCmd.Flags().IntVar(&maxConcurrentProbes, "max-concurrent-probes", 10, "Certprobe worker pool size")
 }
 
 func envOrDefault(key, fallback string) string {
 	if v, ok := os.LookupEnv(key); ok {
 		return v
-	}
-	return fallback
-}
-
-func envDurationOrDefault(key string, fallback time.Duration) time.Duration {
-	if v, ok := os.LookupEnv(key); ok {
-		if d, err := time.ParseDuration(v); err == nil {
-			return d
-		}
 	}
 	return fallback
 }
